requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
djangorestframework==3.12.4
idna==2.10
pyaes==1.6.1
pyasn1==0.4.8
pytz==2021.1
requests==2.25.1
rsa==4.7.2
sqlparse==0.4.1
urllib3==1.26.6
